+++ src/main.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
-import App from './App.jsx'
+import { SimpleRouter } from './router/SimpleRouter.jsx'
 import { LocaleProvider } from './i18n/LocaleProvider'
 import { ThemeProvider } from './contexts/ThemeContext'
 
@@ -9,7 +9,7 @@ createRoot(document.getElementById('root')).render(
   <StrictMode>
     <ThemeProvider>
       <LocaleProvider>
-        <App />
+        <SimpleRouter />
       </LocaleProvider>
     </ThemeProvider>
   </StrictMode>,